third change

+++ hotelProject/hotelProject/Controllers/CustomersController.cs
@@ -24,37 +24,49 @@ namespace hotelProject.Controllers
 
         // GET api/<CustomersController>/5
         [HttpGet("{id}")]
-        public Customers Get(string id)
+        public ActionResult<Customers> Get(string id)
         {
-            return context.customersList.Find(x => x.CustId ==id);
+            Customers c=context.customersList.Find(x => x.CustId ==id);
+            if(c==null)
+                return NotFound();
+            return c;
         }
 
         // POST api/<CustomersController>
         [HttpPost]
-        public void Post([FromBody] Customers c)
+        public ActionResult Post([FromBody] Customers c)
         {
             Customers customers = new Customers { CustId=c.CustId, Name=c.Name, Email=c.Email, Adress=c.Adress,Phone=c.Phone };
+            if (customers != null)
+                return NotFound();
             context.customersList.Add(customers);
+            return Ok();
         }
 
         // PUT api/<CustomersController>/5
         [HttpPut("{id}")]
-        public void Put(string id, [FromBody] Customers c)
+        public ActionResult Put(string id, [FromBody] Customers c)
         {
             Customers cust = context.customersList.Find(x => x.CustId==id);
+            if (cust != null)
+                return NotFound();
             cust.Name = c.Name;
             cust.Email = c.Email;
             cust.Adress = c.Adress;
             cust.Phone = c.Phone;
+            return Ok();
 
         }
 
         // DELETE api/<CustomersController>/5
         [HttpDelete("{id}")]
-        public void Delete(string id)
+        public ActionResult Delete(string id)
         {
              Customers cust = context.customersList.Find(x => x.CustId == id);
+            if (cust != null)
+                return NotFound();
             context.customersList.Remove(cust);
+            return Ok();
          }
     }
 }
