+++ hotelProject/hotelProject/Controllers/RoomsController.cs
@@ -26,36 +26,49 @@ namespace hotelProject.Controllers
 
         // GET api/<RoomsController>/5
         [HttpGet("{id}")]
-        public Rooms Get(int id)
+        public ActionResult< Rooms> Get(int id)
         {
-            return context.roomsList.Find(x=>x.RoomId==id);
+            Rooms r = context.roomsList.Find(x => x.RoomId == id);
+            if (r == null)
+                return NotFound();
+
+            return r;
         }
 
         // POST api/<RoomsController>
         [HttpPost]
-        public void Post([FromBody] Rooms r)
+        public ActionResult Post([FromBody] Rooms r)
         {
             Rooms room = new Rooms { RoomId = r.Floor * 100 + context.roomNumber[r.Floor],  Floor = r.Floor, NumOfBad = r.NumOfBad, Price = r.Price };
+            if (room == null)
+                return NotFound();
             context.roomNumber[r.Floor]++;
             context.roomsList.Add(room);
+            return Ok();
         }
 
         // PUT api/<RoomsController>/5
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Rooms r)
+        public ActionResult Put(int id, [FromBody] Rooms r)
         {
-            Rooms room= context.roomsList.Find(x => x.RoomId == id);   
+            Rooms room= context.roomsList.Find(x => x.RoomId == id);
+            if (room == null)
+                return NotFound();
             room.Floor = r.Floor;
             room.NumOfBad = r.NumOfBad;          
             room.Price = r.Price;
+            return Ok();   
         }
 
         // DELETE api/<RoomsController>/5
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public ActionResult Delete(int id)
         {
             Rooms room = context.roomsList.Find(x => x.RoomId == id);
+            if (room == null)
+                return NotFound();
             context.roomsList.Remove(room);
+            return Ok();
         }
     }
 }
